refactored from danilospa comments
