+++ src/components/Login/index.scss
@@ -33,11 +33,11 @@ div {
   border-radius: 1rem;  
   box-shadow: 3px 6px 50px -4px rgba(0,0,0,0.42);
 
-    .input {
-      width: 500px;
-    }
+  .input {
+    width: 500px;
+  }
 
-    .button {
-      margin-top: 2rem;
-    }
+  .button {
+    margin-top: 2rem;
+  }
 }
